requirements.txt
appdirs==1.4.3
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
convertdate==2.3.2
decorator==5.0.7
distlib==0.3.0
distro==1.4.0
hijri-converter==2.1.1
holidays==0.11.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
korean-lunar-calendar==0.2.1
lockfile==0.12.2
msgpack==0.6.2
numpy==1.19.5
packaging==20.3
pandas==1.1.5
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
PyMeeus==0.5.11
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
tornado==6.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
